+++ src/main.h
@@ -4,8 +4,8 @@
 
 # include "base.h"
 
-void startscreen(SDL_Window *screen,uint *state,uint *grapset,uint *fullscreen);
-void history(SDL_Window *screen,uint *state,uint *grapset,uint *fullscreen);
-void game(SDL_Window *screen,uint *state,uint *grapset,uint *fullscreen);
-void gameover(SDL_Window *screen,uint *state);
-void ending (SDL_Window *screen,uint *state);
+void startscreen(SDL_Window *screen,uint8_t *state,uint8_t *grapset,uint8_t *fullscreen);
+void history(SDL_Window *screen,uint8_t *state,uint8_t *grapset,uint8_t *fullscreen);
+void game(SDL_Window *screen,uint8_t *state,uint8_t *grapset,uint8_t *fullscreen);
+void gameover(SDL_Window *screen,uint8_t *state);
+void ending (SDL_Window *screen,uint8_t *state);
